+++ RELEASE_NOTES.md
@@ -1,11 +1,12 @@
 # Release Notes
 
-## OCD-1845
-_Delete deprecated uploads_
+## Version TBD
+_Date TBD_
 
 ### Data changes
 * Mark deprecated upload templates as deleted
 * Update "new" 2014 upload template to have correct "available_as_of_date"
+* Add new view to use for faster API searches
 
 ---